web-app/requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==2.2.2
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.8
Werkzeug==0.14.1
